+++ packages/ui/libs/ui/theme/src/config/index.ts
@@ -1,0 +1,2 @@
+export * from './themeConfig';
+export * from './variables';
